requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bottle==0.13.4
cffi==2.0.0
click==8.3.1
clr_loader==0.2.10
colorama==0.4.6
fastapi==0.135.2
h11==0.16.0
idna==3.11
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pefile==2024.8.26
pillow==12.1.1
proxy_tools==0.1.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
python-multipart==0.0.22
pythonnet==3.0.5
pywebview==6.1
pywin32-ctypes==0.2.3
setuptools==82.0.1
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
yt-dlp==2026.3.17
